(PC-12472) refactor: move test outside of __init__ file
